b_hobj: cast the scaled mass to float; run_suite: mark merged rows in the table
Two findings from the TODO_CLARA cluster sweep (allocation 193102, h200-02, MODE=oglite).

1. B-HOBJ CRASHED THE PERTURBATION SWEEP -- the only failure in 48 cells.

       link.mass = min(base["mass"][name] * s_mass, MASS_CLIP_KG)
         File "omnigibson/prims/rigid_dynamic_prim.py", line 233, in mass
           self._rigid_prim_view.set_masses(th.tensor([mass]))
       RuntimeError: Index put requires the source and destination dtypes match,
                     got Float for the destination and Double for the source.
       Fatal Python error: Segmentation fault

   `base["mass"][name] * s_mass` is an np.float64 (s_mass comes from np.random), and min() returns
   whichever OPERAND is smaller KEEPING ITS TYPE: below the 2 kg clip it hands back the numpy scalar,
   at the clip it hands back the Python MASS_CLIP_KG. So the written dtype depends on the DRAW and
   only the under-clip branch is wrong -- which for these objects is the usual one. Confirmed both
   branches directly: torch.float64 under the clip, torch.float32 when clipped.

   The joint setters three lines below ALREADY wrap every write in float(), with a comment naming this
   exact hazard. The mass line one block up was missed. Fixed the same way.

   NO NUMBER MOVES: Isaac stores masses as float32 either way, so this only stops the crash.
   Verified: task 0 x B-HOBJ x 3 repeats now runs clean (rc=0, 3/3 report rows, all four artifacts,
   no dtype error), and t10_bhobj_props PASSES on both paths -- task 0 (6 resets) and task 8, the
   joint path (4 resets) -- "every property stays within ONE draw of its baseline (no compounding)
   and still varies from reset to reset (not frozen)". That also gives the 1.0.0 B-HOBJ batch its
   first live confirmation, including the max_velocity/friction setters TODO_CLARA flagged as
   "verified against OmniGibson 3.9.1 source, not live". They write correctly.

   Why the offline equivalence harness could not have caught it: the write only happens against a
   live Isaac, so no host-side check reaches this line.

2. run_suite's TABLE conflated this run with merged rows.

   Gating was always right -- `mine` decides the exit status and the JUnit XML, so a stale row can
   neither pass nor fail a run. But print_table printed merged rows indistinguishably and counted
   them, so `make check` (two container-free tests) printed a 9-row table ending `PASS=9` including
   `test_perturbations_integrity PASS 48 cells` from an earlier invocation against the same --out.
   That stale row contradicted the FAIL measured against the same tree twenty minutes earlier and
   read as fresh evidence the GPU tier was green; I nearly reported it as such.

   Merged rows are now prefixed `*` and the counts split:

       this run: PASS=2
       * merged from earlier runs against this --out, NOT run now: PASS=7

   --report still marks nothing, since then nothing was run. In a suite whose first rule is that
   verdicts come from printed lines and never from exit codes, a printed line that silently mixes
   runs is the wrong failure mode.

Gated on `make check` (ruff clean, tier-1 tests green, 37 host pytest passed) in the same command as
the commit, not beside it.

+++ realm/environments/perturbations/b_hobj.py
@@ -82,7 +82,21 @@ def b_hobj(env: "RealmEnvironmentDynamic") -> None:
         base = _baselines(env, obj)
 
         for name, link in obj._links.items():
-            link.mass = min(base["mass"][name] * s_mass, MASS_CLIP_KG)  # clip at 2.0kg payload
+            # float() for the same reason as the joint setters below -- and it is NOT redundant here
+            # even though min() looks like it returns a plain float. `base["mass"][name] * s_mass` is
+            # an np.float64 (s_mass comes from np.random), and min() returns whichever OPERAND is
+            # smaller, keeping its type: below the clip it hands back the np.float64, at the clip it
+            # hands back the Python MASS_CLIP_KG. So the dtype of what gets written depends on the
+            # DRAW, and only the under-clip branch is wrong.
+            #
+            # Uncast, that branch reaches RigidDynamicPrim.mass -> th.tensor([np.float64]) -> a float64
+            # tensor -> set_masses' `dst[indices] = src` against a float32 destination:
+            #   RuntimeError: Index put requires the source and destination dtypes match,
+            #                 got Float for the destination and Double for the source.
+            # then a SIGSEGV at teardown. Measured 2026-08-19, first live B-HOBJ run after the 1.0.0
+            # batch (logs/todo_clara/item3_logs); object masses here sit well under 2 kg, so the bad
+            # branch is the usual one. No number moves: Isaac stores masses as float32 either way.
+            link.mass = float(min(base["mass"][name] * s_mass, MASS_CLIP_KG))  # clip at 2.0kg payload
 
         for name, joint in obj.joints.items():
             joint: JointPrim
